+++ openpyxl2/chartsheet/chartsheet.py
@@ -1,22 +1,26 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
-from openpyxl2.descriptors import Typed
+from weakref import ref
+
+from openpyxl2.descriptors import Typed, Set
 from openpyxl2.descriptors.excel import ExtensionList
 from openpyxl2.descriptors.serialisable import Serialisable
+from openpyxl2.drawing.spreadsheet_drawing import AbsoluteAnchor
 from openpyxl2.worksheet.page import (
     PageMargins,
     PrintPageSetup
 )
-from openpyxl2.packaging.relationship import Relationship
+from openpyxl2.packaging.relationship import Relationship, RelationshipList
 from openpyxl2.worksheet.drawing import Drawing
 from openpyxl2.worksheet.header_footer import HeaderFooter
 from openpyxl2.workbook.child import _WorkbookChild
+from openpyxl2.xml.constants import SHEET_MAIN_NS
 
 from .relation import DrawingHF, SheetBackgroundPicture
 from .properties import ChartsheetProperties
 from .protection import ChartsheetProtection
-from .views import ChartsheetViews
+from .views import ChartsheetViewList
 from .custom import CustomChartsheetViews
 from .publish import WebPublishItems
 
@@ -27,7 +31,7 @@ class Chartsheet(_WorkbookChild, Serialisable):
     _default_title = "Chart"
 
     sheetPr = Typed(expected_type=ChartsheetProperties, allow_none=True)
-    sheetViews = Typed(expected_type=ChartsheetViews, allow_none=True)
+    sheetViews = Typed(expected_type=ChartsheetViewList)
     sheetProtection = Typed(expected_type=ChartsheetProtection, allow_none=True)
     customSheetViews = Typed(expected_type=CustomChartsheetViews, allow_none=True)
     pageMargins = Typed(expected_type=PageMargins, allow_none=True)
@@ -38,11 +42,14 @@ class Chartsheet(_WorkbookChild, Serialisable):
     picture = Typed(expected_type=SheetBackgroundPicture, allow_none=True)
     webPublishItems = Typed(expected_type=WebPublishItems, allow_none=True)
     extLst = Typed(expected_type=ExtensionList, allow_none=True)
+    sheet_state = Set(values=('visible', 'hidden', 'veryHidden'))
 
     __elements__ = (
         'sheetPr', 'sheetViews', 'sheetProtection', 'customSheetViews',
         'pageMargins', 'pageSetup', 'headerFooter', 'drawing', 'drawingHF',
         'picture', 'webPublishItems')
+
+    __attrs__ = ()
 
     def __init__(self,
                  sheetPr=None,
@@ -59,34 +66,35 @@ class Chartsheet(_WorkbookChild, Serialisable):
                  extLst=None,
                  parent=None,
                  title="",
+                 sheet_state='visible',
                  ):
         # hack to simplify testing
         if parent is not None:
             super(Chartsheet, self).__init__(parent, title)
         self._charts = []
         self.sheetPr = sheetPr
+        if sheetViews is None:
+            sheetViews = ChartsheetViewList()
         self.sheetViews = sheetViews
         self.sheetProtection = sheetProtection
         self.customSheetViews = customSheetViews
         self.pageMargins = pageMargins
         self.pageSetup = pageSetup
         self.headerFooter = headerFooter
-        self.drawing = drawing
+        self.drawing = Drawing("rId1")
         self.drawingHF = drawingHF
         self.picture = picture
         self.webPublishItems = webPublishItems
+        self.sheet_state = sheet_state
 
 
     def add_chart(self, chart):
+        chart.anchor = AbsoluteAnchor()
         self._charts.append(chart)
-        self.parent._charts.append(chart)
+        self.parent._charts.append(ref(chart))
 
 
     def to_tree(self):
-        self._rels = []
-        if self._charts:
-            rel = Relationship(type="drawing", target="")
-            self._rels.append(rel)
-            self.drawing = Drawing()
-            self.drawing.id = "rId%s" % len(self._rels)
-        return super(Chartsheet, self).to_tree()
+        tree = super(Chartsheet, self).to_tree()
+        tree.set("xmlns", SHEET_MAIN_NS)
+        return tree