+++ openpyxl2/chartsheet/tests/test_chartsheet.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 
 import pytest
 from openpyxl2.chartsheet import Drawing
-from ..views import ChartsheetView, ChartsheetViews
+from ..views import ChartsheetView, ChartsheetViewList
 from openpyxl2.worksheet import PageMargins
 
 from openpyxl2.xml.functions import fromstring, tostring
@@ -48,12 +48,13 @@ class TestChartsheet:
     def test_write(self, Chartsheet):
 
         sheetview = ChartsheetView(tabSelected=True, zoomScale=80, workbookViewId=0, zoomToFit=True)
-        chartsheetViews = ChartsheetViews(sheetView=[sheetview])
+        chartsheetViews = ChartsheetViewList(sheetView=[sheetview])
         pageMargins = PageMargins(left=0.7, right=0.7, top=0.75, bottom=0.75, header=0.3, footer=0.3)
         drawing = Drawing("rId1")
         item = Chartsheet(sheetViews=chartsheetViews, pageMargins=pageMargins, drawing=drawing)
         expected = """
-        <chartsheet xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
+        <chartsheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"
+           xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
             <sheetViews>
                 <sheetView tabSelected="1" zoomScale="80" workbookViewId="0" zoomToFit="1"/>
             </sheetViews>
@@ -68,10 +69,18 @@ class TestChartsheet:
 
     def test_write_charts(self, Chartsheet):
 
+        class DummyChart:
+
+            pass
+
         cs = Chartsheet(parent=DummyWorkbook())
-        cs.add_chart(1)
+        cs.add_chart(DummyChart())
         expected = """
-        <chartsheet xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
+        <chartsheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"
+           xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">
+           <sheetViews>
+             <sheetView workbookViewId="0"></sheetView>
+            </sheetViews>
            <drawing r:id="rId1" />
         </chartsheet>
         """