+++ openpyxl2/writer/relations.py
@@ -1,18 +1,13 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2015 openpyxl
 
-
-from openpyxl2.xml.functions import Element
-from openpyxl2.xml.constants import (
-    PKG_REL_NS,
-)
-from openpyxl2.packaging.relationship import Relationship
+from openpyxl2.packaging.relationship import Relationship, RelationshipList
 
 
 def write_rels(worksheet, comments_id=None, vba_controls_id=None):
     """Write relationships for the worksheet to xml."""
-    root = Element('Relationships', xmlns=PKG_REL_NS)
-    rels = worksheet._rels
+
+    rels = RelationshipList(worksheet._rels)
 
     # VBA
     if worksheet.vba_controls is not None:
@@ -31,9 +26,4 @@ def write_rels(worksheet, comments_id=None, vba_controls_id=None):
                            target='/xl/drawings/commentsDrawing%s.vml' % comments_id)
             rels.append(rel)
 
-    for idx, rel in enumerate(rels, 1):
-        if rel.id is None:
-            rel.id = "rId{0}".format(idx)
-        root.append(rel.to_tree())
-
-    return root
+    return rels.to_tree()